backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
google-genai>=1.0.0
qdrant-client>=1.7.1
gitpython>=3.1.41
pymupdf>=1.23.20
beautifulsoup4>=4.12.3
httpx>=0.26.0
python-multipart>=0.0.6
pydantic>=2.5.3
aiofiles>=23.2.1
google-auth>=2.27.0
python-jose[cryptography]>=3.3.0
groq>=0.4.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.26.0
anthropic>=0.18.0
openai>=1.12.0
tavily-python>=0.3.0
google-api-python-client>=2.100.0
slowapi>=0.1.9
sentry-sdk[fastapi]>=1.40.0
redis>=5.0.0
cross-encoder>=0.3.0
